small fixes
