+++ config/vufind/ComposedDriver.ini
@@ -8,7 +8,7 @@
 ;drivers_config_path = private
 
 ; Definition of all the used drivers in the form <DriverName> = <DriverClass>
-; The name of config file for a driver needs te be the same <DriverName>
+; The name of config file for a driver needs to be the same <DriverName>
 [Drivers]
 ;driver1 = PAIA
 ;driver2 = DAIA
@@ -22,8 +22,9 @@
 ; The array support_keys specifies which of there keys of the given support driver should be composed with the main result.
 ; The format is: support_keys["<DriverName"] = "<key1>,<key2>,...,<keyN>"
 
-; Here are some examples how a configuration could look like for a combination of DAIA and PAIA
-; but it most likely wont work with the Demo driver
+; Here are some examples of how a configuration could look like for a combination of DAIA and PAIA
+; but 'getHolding' most likely won't work with driver3. The Demo driver does not provide the same item_ids
+; as the other drivers.
 ;[getStatus]
 ;main_driver = "driver2"
 ;
