+++ module/VuFind/src/VuFind/ILS/Driver/AbstractMultiDriverFactory.php
@@ -46,7 +46,7 @@ use Psr\Container\ContainerInterface;
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development Wiki
  */
-class ComposedDriverFactory implements FactoryInterface
+class AbstractMultiDriverFactory implements FactoryInterface
 {
     /**
      * Create an object
