+++ module/VuFind/src/VuFind/ILS/Driver/ComposedDriver.php
@@ -59,20 +59,6 @@ class ComposedDriver extends AbstractMultiDriver
     protected $mainDriver;
 
     /**
-     * Constructor
-     *
-     * @param \VuFind\Config\PluginManager $configLoader Configuration loader
-     * @param PluginManager                $dm           ILS driver manager
-     */
-    public function __construct(
-        \VuFind\Config\PluginManager $configLoader,
-        PluginManager $dm
-    ) {
-        $this->configLoader = $configLoader;
-        $this->driverManager = $dm;
-    }
-
-    /**
      * Initialize the driver.
      *
      * Validate configuration and perform all resource-intensive tasks needed to
