+++ module/VuFind/tests/unit-tests/src/VuFindTest/ILS/Driver/MultiBackendTest.php
@@ -32,6 +32,7 @@
 namespace VuFindTest\ILS\Driver;
 
 use Laminas\Config\Exception\RuntimeException;
+use VuFind\Exception\ILS as ILSException;
 use VuFind\ILS\Driver\MultiBackend;
 
 use function call_user_func_array;
@@ -2554,5 +2555,46 @@ class MultiBackendTest extends AbstractMultiDriverTest
             $constructorArgs['driverManager'] ?? $this->getMockSM()
         );
         return $driver;
+    }
+
+    /**
+     * Get a mock ILS authenticator
+     *
+     * @param string $userSource Source id, if the authenticator should emulate a
+     * situation where a user has logged in. Set to null for the attempt to cause an
+     * exception.
+     *
+     * @return \VuFind\Auth\ILSAuthenticator
+     */
+    protected function getMockILSAuthenticator($userSource = '')
+    {
+        $mockAuth = $this->getMockBuilder(\VuFind\Auth\ILSAuthenticator::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+        if ($userSource) {
+            $mockAuth->expects($this->any())
+                ->method('storedCatalogLogin')
+                ->will(
+                    $this->returnValue($this->getPatron('username', $userSource))
+                );
+            $mockAuth->expects($this->any())
+                ->method('getStoredCatalogCredentials')
+                ->will(
+                    $this->returnValue($this->getPatron('username', $userSource))
+                );
+        } elseif (null === $userSource) {
+            $e = new ILSException('Simulated exception from ILSAuthenticator');
+            $mockAuth->expects($this->any())
+                ->method('storedCatalogLogin')
+                ->will(
+                    $this->throwException($e)
+                );
+            $mockAuth->expects($this->any())
+                ->method('getStoredCatalogCredentials')
+                ->will(
+                    $this->throwException($e)
+                );
+        }
+        return $mockAuth;
     }
 }
